requirements.txt
python-dotenv
google-cloud-vision
gunicorn
pytest
pytest-cov